backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0
channels==4.0.0
celery==5.3.1
redis==4.5.5
psycopg2-binary==2.9.6
Pillow==9.5.0
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
model-bakery==1.12.0
drf-spectacular==0.26.4
gunicorn==20.1.0
whitenoise==6.5.0
